进行到遍历ID，正要写读取初始化状态，然后要把这里的全局变量加个结构体，用this包起来

+++ cycleflash/include/cycle_flash_system.h
@@ -41,7 +41,7 @@
 cfs_object_handle_ptr cfs_nv_object_init(
     const uint8_t *name, 
     const uint32_t address,
-    const uint16_t sector_count,
+    const uint32_t sector_count,
     const uint16_t data_size);
