+++ cycleflash/src/cycle_flash_system.c
@@ -305,7 +305,6 @@ static uint32_t cfs_filesystem_flsh_data_read( \
 
 
 
-
 //*******************************************************************************************
 //-- 对外接口  
 //*******************************************************************************************
@@ -314,7 +313,7 @@ static uint32_t cfs_filesystem_flsh_data_read( \
 cfs_object_handle_ptr cfs_nv_object_init(
     const uint8_t *name, 
     const uint32_t address,
-    const uint16_t sector_count, 
+    const uint32_t sector_count, 
     const uint16_t data_size)
 {
     //@def 判断参数有效性
